refactor: `Tab`の階層を移動
